graphコンポーネントのuseEffectをhooksにまとめる
